ENG-27826: KNI api to set interface stats
KNI keeps interface stats it exposes to Linux network stack.
Add an api to set the values of these counters from the userspace owner
of the kni interface - so that they could reflect the values tracked for
the corresponding physical interface.

There is some fudge involved because some of these counters are updated
by the driver itself, so there could be temporary deviations from the
set values - although if all kni packets are also accounted on the
'master' interface  the values should not diverge much.