VLENG-44508: i40evf: async eth stats
i40e implements access to VF stats via async call to the PF over the
'admin queue'. The driver implements it call as a blocking busy-polling
loop waiting for the interrupt thread to indicate the completion - both
inefficiently resource consuming and prone to priority inversion the
waiting thread can starve the completion thread out of resources and
prevent or delay the completion.

Linux driver implements this function more reasonably by keeping the
last known stats available for read without waiting, a background task
to perform stats queries and a background task to poll the admin queue.

This is a simplified implementation combining some of the same together
- the get_stats API returnes the last known stats and at the same time
initiates the PF query, the completion handler updates the last known
stats when the response comes back. This is far from perfect but avoids
the busy loops while keeping the stats reasonably current.