VLENG-53788: linux 5.4 compilation fix
In linux 5.4, linux/platform_data/mdio-gpio.h has been stripped down
to almost nothing.  But what we really needed from there was just
linux/mdio-bitbang.h (which used to be included from that file), so
we might as well just include that directly instead (and that works
on all linux versions going back to 3.x).